Moved Climb-Up/Down plugin into MenuEntrySwapper
