+++ runelite-client/src/main/java/net/runelite/client/plugins/menuentryswapper/MenuEntrySwapperConfig.java
@@ -302,4 +302,14 @@ public interface MenuEntrySwapperConfig extends Config
 	{
 		return true;
 	}
+
+	@ConfigItem(
+			keyName = "swapClimbUpDown",
+			name = "Climb",
+			description = "Swap Climb-Up/Down depending on Shift or Control key									"
+	)
+	default boolean swapClimbUpDown()
+	{
+		return false;
+	}
 }
